refac(client): kahos types structure
